refactor(analysis): isolate box enclosure contracts
