+++ src/jacobian/math/analysis/_operations.py
@@ -8,21 +8,15 @@ from typing import Any
 
 from jacobian.catalog._examples import example
 from jacobian.catalog.models import MathTool
-from jacobian.math.analysis._arb import dyadic_endpoints
+from jacobian.math.analysis._arb import arb_source_interval, dyadic_endpoints
 from jacobian.math.analysis._expression_enclosure import (
     IntervalExpressionEnclosureRequest,
     IntervalExpressionEnclosureResult,
 )
 from jacobian.math.analysis._models import (
-    MAX_BOX_PREFLIGHT_TEMPORARY_BITS,
     DyadicClosedInterval,
-    ExactDyadic,
-    IntervalExpressionBoxEnclosureRequest,
-    IntervalExpressionBoxEnclosureResult,
-    IntervalExpressionBoxEnclosureStatus,
     IntervalExpressionDomainFailure,
     IntervalExpressionNode,
-    _preflight_box_expression,
     _rational_box_bounds,
 )
 from jacobian.math.analysis._point_enclosure import (
@@ -41,7 +35,6 @@ from jacobian.math.analysis._second_jet import (
     IntervalExpressionSecondJetEnclosureStatus,
     _preflight_second_jet_expression,
 )
-from jacobian.math.geometry.boxes.values import RationalClosedInterval
 
 
 class _EvaluationFailure(StrEnum):
@@ -173,277 +166,6 @@ def _expression_enclosure(
         relative_accuracy_bits=None if exact else int(result.rel_accuracy_bits()),
         exact=exact,
         detail="Arb returned an outward-rounded enclosure with exact dyadic endpoints.",
-    )
-
-
-def _normalize_dyadic_pair(mantissa: int, exponent: int) -> tuple[int, int]:
-    if mantissa == 0:
-        return 0, 0
-    while mantissa % 2 == 0:
-        mantissa //= 2
-        exponent += 1
-    return mantissa, exponent
-
-
-def _add_dyadic_pairs(left: tuple[int, int], right: tuple[int, int]) -> tuple[int, int]:
-    common_exponent = min(left[1], right[1])
-    return _normalize_dyadic_pair(
-        (left[0] << (left[1] - common_exponent))
-        + (right[0] << (right[1] - common_exponent)),
-        common_exponent,
-    )
-
-
-def _negate_dyadic_pair(value: tuple[int, int]) -> tuple[int, int]:
-    return -value[0], value[1]
-
-
-def _halve_dyadic_pair(value: tuple[int, int]) -> tuple[int, int]:
-    return _normalize_dyadic_pair(value[0], value[1] - 1)
-
-
-def _exact_arb_dyadic_pair(value: Any) -> tuple[int, int]:
-    mantissa, exponent = value.man_exp()
-    return _normalize_dyadic_pair(int(mantissa), int(exponent))
-
-
-def _arb_source_interval(interval: RationalClosedInterval) -> Any:
-    """Build one Arb ball that contains the exact rational source interval.
-
-    Arb radii have a fixed implementation precision.  Anchoring a one-sided
-    interval at its endpoint nearest zero preserves a proved sign even when
-    the radius is rounded upward by python-flint's public constructor.
-    """
-
-    from flint import arb, fmpq
-
-    lower_ratio = interval.lower.as_integer_ratio()
-    upper_ratio = interval.upper.as_integer_ratio()
-    if lower_ratio == upper_ratio:
-        return arb(fmpq(*lower_ratio))
-
-    lower = _exact_arb_dyadic_pair(arb(fmpq(*lower_ratio)).lower())
-    upper = _exact_arb_dyadic_pair(arb(fmpq(*upper_ratio)).upper())
-    half_width = _halve_dyadic_pair(
-        _add_dyadic_pairs(upper, _negate_dyadic_pair(lower))
-    )
-    actual_radius = _exact_arb_dyadic_pair(arb((0, 0), half_width).upper())
-    if interval.lower.as_fraction() >= 0:
-        midpoint = _add_dyadic_pairs(lower, actual_radius)
-    elif interval.upper.as_fraction() <= 0:
-        midpoint = _add_dyadic_pairs(upper, _negate_dyadic_pair(actual_radius))
-    else:
-        midpoint = _halve_dyadic_pair(_add_dyadic_pairs(lower, upper))
-    return arb(midpoint, half_width)
-
-
-def _box_domain_failure(
-    node: IntervalExpressionNode, path: tuple[int, ...]
-) -> IntervalExpressionDomainFailure:
-    if node.op == "div":
-        return IntervalExpressionDomainFailure(
-            node_path=path,
-            operation="div",
-            reason="DENOMINATOR_CONTAINS_ZERO",
-        )
-    if node.op == "pow":
-        return IntervalExpressionDomainFailure(
-            node_path=path,
-            operation="pow",
-            reason="NEGATIVE_POWER_BASE_CONTAINS_ZERO",
-        )
-    if node.op == "log":
-        return IntervalExpressionDomainFailure(
-            node_path=path,
-            operation="log",
-            reason="LOG_ARGUMENT_NOT_STRICTLY_POSITIVE",
-        )
-    if node.op == "sqrt":
-        return IntervalExpressionDomainFailure(
-            node_path=path,
-            operation="sqrt",
-            reason="SQRT_ARGUMENT_NOT_NONNEGATIVE",
-        )
-    raise AssertionError("only real-domain operations can produce a domain failure")
-
-
-def _evaluate_box_unary(
-    node: IntervalExpressionNode, value: Any, path: tuple[int, ...]
-) -> Any:
-    if node.op == "neg":
-        return -value
-    if node.op == "pow":
-        assert node.exponent is not None
-        if node.exponent < 0 and value.contains(0):
-            return _box_domain_failure(node, path)
-        return value**node.exponent
-    if node.op == "log":
-        if not value > 0:
-            return _box_domain_failure(node, path)
-        return value.log()
-    if node.op == "sqrt":
-        if not value >= 0:
-            return _box_domain_failure(node, path)
-        return value.sqrt()
-    if node.op in ("exp", "sin", "cos"):
-        return getattr(value, node.op)()
-    raise AssertionError(f"unsupported unary expression operation: {node.op}")
-
-
-def _evaluate_box_binary(
-    node: IntervalExpressionNode,
-    left: Any,
-    right: Any,
-    path: tuple[int, ...],
-) -> Any:
-    if node.op == "add":
-        return left + right
-    if node.op == "sub":
-        return left - right
-    if node.op == "mul":
-        return left * right
-    if node.op == "div":
-        if right.contains(0):
-            return _box_domain_failure(node, path)
-        return left / right
-    raise AssertionError(f"unsupported binary expression operation: {node.op}")
-
-
-def _evaluate_box_expression(
-    node: IntervalExpressionNode,
-    variables: dict[str, Any],
-    path: tuple[int, ...] = (),
-) -> Any:
-    from flint import arb, fmpq
-
-    if node.op == "const":
-        assert node.value is not None
-        return arb(fmpq(*node.value.as_integer_ratio()))
-    if node.op == "var":
-        assert node.variable is not None
-        return variables[node.variable]
-
-    values: list[Any] = []
-    for index, child in enumerate(node.children):
-        value = _evaluate_box_expression(child, variables, (*path, index))
-        if isinstance(value, (IntervalExpressionDomainFailure, _BoxEvaluationFailure)):
-            return value
-        if not value.is_finite():
-            return _BoxEvaluationFailure.BACKEND_ERROR
-        values.append(value)
-
-    left = values[0]
-    if len(values) == 1:
-        return _evaluate_box_unary(node, left, path)
-
-    right = values[1]
-    return _evaluate_box_binary(node, left, right, path)
-
-
-def _constructed_box_result(
-    request: IntervalExpressionBoxEnclosureRequest,
-    *,
-    status: IntervalExpressionBoxEnclosureStatus,
-    detail: str,
-    lower: ExactDyadic | None = None,
-    upper: ExactDyadic | None = None,
-    domain_failure: IntervalExpressionDomainFailure | None = None,
-) -> IntervalExpressionBoxEnclosureResult:
-    """Build the producer result without paying a second backend replay."""
-
-    return IntervalExpressionBoxEnclosureResult.model_construct(
-        expression=request.expression,
-        box=request.box,
-        precision_bits=request.precision_bits,
-        status=status,
-        lower=lower,
-        upper=upper,
-        domain_failure=domain_failure,
-        method="ARB_NATURAL_INTERVAL_EXTENSION",
-        detail=detail,
-    )
-
-
-def _box_expression_enclosure(
-    request: IntervalExpressionBoxEnclosureRequest,
-) -> IntervalExpressionBoxEnclosureResult:
-    preflight = _preflight_box_expression(
-        request.expression, _rational_box_bounds(request.box)
-    )
-    if isinstance(preflight, IntervalExpressionDomainFailure):
-        return _constructed_box_result(
-            request,
-            status="DOMAIN_UNPROVEN",
-            domain_failure=preflight,
-            detail=(
-                "The exact admission interval extension could not establish the "
-                "real domain at the reported source node."
-            ),
-        )
-
-    from flint import ctx
-
-    try:
-        with ctx.workprec(request.precision_bits):
-            variables = {
-                variable: _arb_source_interval(interval)
-                for variable, interval in zip(
-                    request.box.variables, request.box.intervals, strict=True
-                )
-            }
-            result = _evaluate_box_expression(request.expression, variables)
-            if isinstance(result, IntervalExpressionDomainFailure):
-                return _constructed_box_result(
-                    request,
-                    status="DOMAIN_UNPROVEN",
-                    domain_failure=result,
-                    detail=(
-                        "Arb's natural interval extension could not establish the "
-                        "real domain at the reported source node."
-                    ),
-                )
-            if isinstance(result, _BoxEvaluationFailure) or not result.is_finite():
-                return _constructed_box_result(
-                    request,
-                    status="BACKEND_ERROR",
-                    detail=(
-                        "Pinned Arb returned no finite enclosure within the admitted "
-                        "fixed-precision envelope."
-                    ),
-                )
-            lower_mantissa, lower_exponent = result.lower().man_exp()
-            upper_mantissa, upper_exponent = result.upper().man_exp()
-            endpoints = dyadic_endpoints(
-                lower_mantissa, lower_exponent, upper_mantissa, upper_exponent
-            )
-    except (OverflowError, ValueError):
-        return _constructed_box_result(
-            request,
-            status="BACKEND_ERROR",
-            detail=(
-                "Pinned Arb rejected an admitted bounded computation; no enclosure "
-                "conclusion is available."
-            ),
-        )
-
-    if endpoints is None:
-        return _constructed_box_result(
-            request,
-            status="BACKEND_ERROR",
-            detail=(
-                "Pinned Arb produced endpoints outside the admitted dyadic wire "
-                "envelope; no enclosure conclusion is available."
-            ),
-        )
-    return _constructed_box_result(
-        request,
-        status="ENCLOSED",
-        lower=endpoints[0],
-        upper=endpoints[1],
-        detail=(
-            "Pinned Arb natural interval arithmetic returned an outward-rounded "
-            "enclosure with exact dyadic endpoints."
-        ),
     )
 
 
@@ -708,7 +430,7 @@ def _second_jet_enclosure(
 
         with ctx.workprec(request.precision_bits):
             variables = {
-                variable: _arb_source_interval(interval)
+                variable: arb_source_interval(interval)
                 for variable, interval in zip(
                     request.box.variables, request.box.intervals, strict=True
                 )
@@ -901,57 +623,6 @@ EXPRESSION_ENCLOSURE_OPERATIONS = (
     ),
 )
 
-BOX_EXPRESSION_ENCLOSURE_OPERATIONS = (
-    MathTool(
-        operation_id="interval.expression.box_enclosure.compute",
-        title="Enclose an elementary expression over a rational box",
-        description=(
-            "Use pinned Arb natural interval arithmetic to enclose one bounded "
-            "named-variable elementary expression over a complete ordered rational "
-            "box, or report the first source node whose real domain is unproved. "
-            "The fixed envelope admits at most 8 variables, 64 nodes, depth 16, "
-            "128-digit rationals, absolute power exponents up to 64, 4,096-bit "
-            "Arb precision, 8,192-bit retained exact admission bounds, and "
-            f"{MAX_BOX_PREFLIGHT_TEMPORARY_BITS:,}-bit Fraction temporaries."
-        ),
-        request_type=IntervalExpressionBoxEnclosureRequest,
-        result_type=IntervalExpressionBoxEnclosureResult,
-        run=_box_expression_enclosure,
-        tags=(
-            "analysis",
-            "interval",
-            "expression",
-            "box",
-            "multivariate",
-            "arb",
-            "validated",
-            "bounded",
-        ),
-        examples=(
-            example(
-                "exp_unit_interval",
-                "Enclose exp(x) over the exact rational interval 0 <= x <= 1.",
-                {
-                    "expression": {
-                        "op": "exp",
-                        "children": [{"op": "var", "variable": "x"}],
-                    },
-                    "box": {
-                        "variables": ["x"],
-                        "intervals": [
-                            {
-                                "lower": {"num": "0", "den": "1"},
-                                "upper": {"num": "1", "den": "1"},
-                            }
-                        ],
-                    },
-                    "precision_bits": 128,
-                },
-            ),
-        ),
-    ),
-)
-
 SECOND_JET_ENCLOSURE_OPERATIONS = (
     MathTool(
         operation_id="interval.expression.second_jet_enclosure.compute",
@@ -1020,7 +691,6 @@ SECOND_JET_ENCLOSURE_OPERATIONS = (
 )
 
 __all__ = [
-    "BOX_EXPRESSION_ENCLOSURE_OPERATIONS",
     "EXPRESSION_ENCLOSURE_OPERATIONS",
     "POINT_ENCLOSURE_OPERATIONS",
     "SECOND_JET_ENCLOSURE_OPERATIONS",
