+++ src/jacobian/math/analysis/_tools.py
@@ -1,8 +1,8 @@
 """Validated real-analysis operations."""
 
 from jacobian.catalog.models import MathTools
+from jacobian.math.analysis._box_enclosure import BOX_EXPRESSION_ENCLOSURE_OPERATIONS
 from jacobian.math.analysis._operations import (
-    BOX_EXPRESSION_ENCLOSURE_OPERATIONS,
     EXPRESSION_ENCLOSURE_OPERATIONS,
     POINT_ENCLOSURE_OPERATIONS,
     SECOND_JET_ENCLOSURE_OPERATIONS,
